Fix a leaking pytest-cov abstraction.
The CoverageReportType enum is pytest-cov specific but lived in the
generic test goal module. Move it to where it belongs in coverage_py.

[ci skip-rust]
[ci skip-build-wheels]
